Reverting rename and moving classes around.
