+++ app/src/androidTest/java/ch/epfl/sweng/rps/StatisticsFragmentTest.kt
@@ -14,13 +14,13 @@ import androidx.test.espresso.util.HumanReadables
 import androidx.test.espresso.util.TreeIterables
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.platform.app.InstrumentationRegistry
-import ch.epfl.sweng.rps.logic.Env
-import ch.epfl.sweng.rps.logic.LocalRepository
+import ch.epfl.sweng.rps.db.Env
+import ch.epfl.sweng.rps.db.LocalRepository
 import ch.epfl.sweng.rps.models.Game
 import ch.epfl.sweng.rps.models.Hand
 import ch.epfl.sweng.rps.models.Round
 import ch.epfl.sweng.rps.models.User
-import ch.epfl.sweng.rps.logic.ServiceLocator
+import ch.epfl.sweng.rps.services.ServiceLocator
 import com.google.firebase.Timestamp
 import org.hamcrest.Matcher
 import org.junit.After
