+++ app/src/main/java/ch/epfl/sweng/rps/db/Env.kt
@@ -1,4 +1,4 @@
-package ch.epfl.sweng.rps.logic
+package ch.epfl.sweng.rps.db
 
 enum class Env(val value: String) {
     Test("test"),
