+++ app/src/main/java/ch/epfl/sweng/rps/ui/statistics/MatchDetailsFragment.kt
@@ -13,7 +13,7 @@ import androidx.core.view.setPadding
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import ch.epfl.sweng.rps.R
-import ch.epfl.sweng.rps.logic.FirebaseHelper
+import ch.epfl.sweng.rps.db.FirebaseHelper
 import kotlinx.coroutines.launch
