+++ app/src/test/java/ch/epfl/sweng/rps/db/UserNotLoggedInTest.kt
@@ -1,6 +1,6 @@
-package ch.epfl.sweng.rps.logic
+package ch.epfl.sweng.rps.db
 
-import ch.epfl.sweng.rps.logic.Repository.UserNotLoggedIn
+import ch.epfl.sweng.rps.db.Repository.UserNotLoggedIn
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
