+++ app/src/test/java/ch/epfl/sweng/rps/db/implementations/FirebaseHelperTest.kt
@@ -1,6 +1,6 @@
-package ch.epfl.sweng.rps.logic.implementations
+package ch.epfl.sweng.rps.db.implementations
 
-import ch.epfl.sweng.rps.logic.FirebaseHelper
+import ch.epfl.sweng.rps.db.FirebaseHelper
 import ch.epfl.sweng.rps.models.User
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
